chore: move qpu list_quantum_processors
