+++ crates/lib/src/api.rs
@@ -1,7 +1,7 @@
 //! This module provides convenience functions to handle compilation,
 //! translation, parameter arithmetic rewriting, and results collection.
 
-use std::{collections::HashMap, time::Duration};
+use std::collections::HashMap;
 
 use num::Complex;
 use qcs_api_client_grpc::{
@@ -10,16 +10,12 @@ use qcs_api_client_grpc::{
         get_controller_job_results_request::Target, GetControllerJobResultsRequest,
     },
 };
-use qcs_api_client_openapi::apis::{
-    quantum_processors_api, translation_api, Error as OpenAPIError,
-};
 
 use serde::Serialize;
-use tokio::time::error::Elapsed;
 
 use crate::qpu::{
     self,
-    client::{GrpcClientError, Qcs, DEFAULT_HTTP_API_TIMEOUT},
+    client::{GrpcClientError, Qcs},
     runner,
 };
 
@@ -140,66 +136,4 @@ pub async fn retrieve_results(
         .result
         .map(ExecutionResults::from)
         .ok_or_else(|| GrpcClientError::ResponseEmpty("Controller Job Execution Results".into()))
-}
-
-/// API Errors encountered when trying to list available quantum processors.
-#[derive(Debug, thiserror::Error)]
-pub enum ListQuantumProcessorsError {
-    /// Failed the http call
-    #[error("Failed to list processors via API: {0}")]
-    ApiError(#[from] OpenAPIError<quantum_processors_api::ListQuantumProcessorsError>),
-
-    /// Pagination did not finish before timeout
-    #[error("API pagination did not finish before timeout.")]
-    TimeoutError(#[from] Elapsed),
-}
-
-/// Query the QCS API for the names of all available quantum processors.
-/// If `None`, the default `timeout` used is 10 seconds.
-pub async fn list_quantum_processors(
-    client: &Qcs,
-    timeout: Option<Duration>,
-) -> Result<Vec<String>, ListQuantumProcessorsError> {
-    let timeout = timeout.unwrap_or(DEFAULT_HTTP_API_TIMEOUT);
-
-    tokio::time::timeout(timeout, async move {
-        let mut quantum_processors = vec![];
-        let mut page_token = None;
-
-        loop {
-            let result = quantum_processors_api::list_quantum_processors(
-                &client.get_openapi_client(),
-                Some(100),
-                page_token.as_deref(),
-            )
-            .await?;
-
-            let mut data = result
-                .quantum_processors
-                .into_iter()
-                .map(|qpu| qpu.id)
-                .collect::<Vec<_>>();
-            quantum_processors.append(&mut data);
-
-            page_token = result.next_page_token;
-            if page_token.is_none() {
-                break;
-            }
-        }
-
-        Ok(quantum_processors)
-    })
-    .await?
-}
-
-/// API Errors encountered when trying to get Quil-T calibrations.
-#[derive(Debug, thiserror::Error)]
-pub enum GetQuiltCalibrationsError {
-    /// Failed the http call
-    #[error("Failed to get Quil-T calibrations via API: {0}")]
-    ApiError(#[from] OpenAPIError<translation_api::GetQuiltCalibrationsError>),
-
-    /// API call did not finish before timeout
-    #[error("API call did not finish before timeout.")]
-    TimeoutError(#[from] Elapsed),
 }
