+++ crates/lib/src/qpu/mod.rs
@@ -1,13 +1,17 @@
 //! This module contains all the functionality for running Quil programs on a real QPU. Specifically,
 //! the [`Execution`] struct in this module.
+use std::time::Duration;
 
-use self::client::OpenApiClientError;
 use qcs_api_client_openapi::{
-    apis::quantum_processors_api::{
-        get_instruction_set_architecture, GetInstructionSetArchitectureError,
+    apis::{
+        quantum_processors_api::{
+            self, get_instruction_set_architecture, GetInstructionSetArchitectureError,
+        },
+        Error as OpenApiError,
     },
     models::InstructionSetArchitecture,
 };
+use tokio::time::error::Elapsed;
 
 pub mod client;
 mod execution;
@@ -36,8 +40,58 @@ pub async fn get_isa(
 ) -> Result<InstructionSetArchitecture, IsaError> {
     get_instruction_set_architecture(&client.get_openapi_client(), quantum_processor_id)
         .await
-        .map_err(OpenApiClientError::RequestFailed)
+        .map_err(client::OpenApiClientError::RequestFailed)
 }
 
 /// Error raised due to failure to get an ISA
-pub type IsaError = OpenApiClientError<GetInstructionSetArchitectureError>;
+pub type IsaError = client::OpenApiClientError<GetInstructionSetArchitectureError>;
+
+/// API Errors encountered when trying to list available quantum processors.
+#[derive(Debug, thiserror::Error)]
+pub enum ListQuantumProcessorsError {
+    /// Failed the http call
+    #[error("Failed to list processors via API: {0}")]
+    ApiError(#[from] OpenApiError<quantum_processors_api::ListQuantumProcessorsError>),
+
+    /// Pagination did not finish before timeout
+    #[error("API pagination did not finish before timeout.")]
+    TimeoutError(#[from] Elapsed),
+}
+
+/// Query the QCS API for the names of all available quantum processors.
+/// If `None`, the default `timeout` used is 10 seconds.
+pub async fn list_quantum_processors(
+    client: &Qcs,
+    timeout: Option<Duration>,
+) -> Result<Vec<String>, ListQuantumProcessorsError> {
+    let timeout = timeout.unwrap_or(client::DEFAULT_HTTP_API_TIMEOUT);
+
+    tokio::time::timeout(timeout, async move {
+        let mut quantum_processors = vec![];
+        let mut page_token = None;
+
+        loop {
+            let result = quantum_processors_api::list_quantum_processors(
+                &client.get_openapi_client(),
+                Some(100),
+                page_token.as_deref(),
+            )
+            .await?;
+
+            let mut data = result
+                .quantum_processors
+                .into_iter()
+                .map(|qpu| qpu.id)
+                .collect::<Vec<_>>();
+            quantum_processors.append(&mut data);
+
+            page_token = result.next_page_token;
+            if page_token.is_none() {
+                break;
+            }
+        }
+
+        Ok(quantum_processors)
+    })
+    .await?
+}
