+++ crates/python/src/api.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 use pyo3::{
     create_exception,
     exceptions::{PyRuntimeError, PyValueError},
@@ -7,11 +5,10 @@ use pyo3::{
     types::{PyComplex, PyFloat, PyInt, PyString},
     Py, PyResult,
 };
-use qcs::api::{self, ExecutionResult, ExecutionResults, Register};
+use qcs::api::{ExecutionResult, ExecutionResults, Register};
 use qcs_api_client_openapi::models::GetQuiltCalibrationsResponse;
 use rigetti_pyo3::{
-    create_init_submodule, impl_repr, py_wrap_data_struct, py_wrap_error, py_wrap_type,
-    py_wrap_union_enum, wrap_error, ToPythonError,
+    create_init_submodule, impl_repr, py_wrap_data_struct, py_wrap_type, py_wrap_union_enum,
 };
 
 use crate::{py_sync::py_function_sync_async, qpu::client::PyQcsClient};
@@ -24,14 +21,11 @@ create_init_submodule! {
     ],
     errors: [
         ExecutionError,
-        DeviceISAError,
-        QCSListQuantumProcessorsError
+        DeviceISAError
     ],
     funcs: [
         py_retrieve_results,
-        py_retrieve_results_async,
-        py_list_quantum_processors,
-        py_list_quantum_processors_async
+        py_retrieve_results_async
     ],
 }
 
@@ -64,26 +58,6 @@ py_wrap_type! {
 create_exception!(qcs, ExecutionError, PyRuntimeError);
 create_exception!(qcs, DeviceISAError, PyValueError);
 
-wrap_error!(ListQuantumProcessorsError(
-    qcs::api::ListQuantumProcessorsError
-));
-py_wrap_error!(
-    api,
-    ListQuantumProcessorsError,
-    QCSListQuantumProcessorsError,
-    PyRuntimeError
-);
-
-wrap_error!(GetQuiltCalibrationsError(
-    qcs::api::GetQuiltCalibrationsError
-));
-py_wrap_error!(
-    api,
-    GetQuiltCalibrationsError,
-    QCSGetQuiltCalibrationsError,
-    PyRuntimeError
-);
-
 py_function_sync_async! {
     #[pyfunction(client = "None")]
     async fn retrieve_results(
@@ -96,20 +70,5 @@ py_function_sync_async! {
             .await
             .map(PyExecutionResults::from)
             .map_err(|e| ExecutionError::new_err(e.to_string()))
-    }
-}
-
-py_function_sync_async! {
-    #[pyfunction(client = "None", timeout = "None")]
-    async fn list_quantum_processors(
-        client: Option<PyQcsClient>,
-        timeout: Option<f64>,
-    ) -> PyResult<Vec<String>> {
-        let client = PyQcsClient::get_or_create_client(client).await?;
-        let timeout = timeout.map(Duration::from_secs_f64);
-        api::list_quantum_processors(&client, timeout)
-            .await
-            .map_err(ListQuantumProcessorsError::from)
-            .map_err(ListQuantumProcessorsError::to_py_err)
     }
 }
